move back to single SettingsActivity
