+++ app/src/main/res/layout/content_settings_protections.xml
